remove GetRequestAWS from ApiUtils, moved to sigv4.cpp
