+++ src/include/api_utils.hpp
@@ -19,8 +19,6 @@ using namespace duckdb_yyjson;
 
 namespace duckdb {
 
-static string SELECTED_CURL_CERT_PATH = "";
-
 // we statically compile in libcurl, which means the cert file location of the build machine is the
 // place curl will look. But not every distro has this file in the same location, so we search a
 // number of common locations and use the first one we find.
@@ -38,11 +36,6 @@ static string certFileLocations[] = {
 
 class APIUtils {
 public:
-	//! We use a global here to store the path that is selected on the ICAPI::InitializeCurl call
-
-	static string GetRequestAws(ClientContext &context, IRCEndpointBuilder endpoint_builder, const string &secret_name);
-	static string GetAwsRegion(const string &host);
-	static string GetAwsService(const string &host);
 	static string GetRequest(ClientContext &context, const IRCEndpointBuilder &endpoint_builder,
 	                         RequestInput &request_input, const string &token = "");
 	static string DeleteRequest(ClientContext &context, const string &url, RequestInput &request_input,
@@ -51,6 +44,8 @@ public:
 	static string PostRequest(ClientContext &context, const string &url, const string &post_data,
 	                          RequestInput &request_input, const string &content_type = "x-www-form-urlencoded",
 	                          const string &token = "");
+	//! We use a singleton here to store the path, set by SelectCurlCertPath
+	static string &GetCURLCertPath();
 };
 
 } // namespace duckdb
